Implement GetAt for Registers
