+++ crates/cpu/src/lib.rs
@@ -7,8 +7,11 @@ pub mod interface;
 pub mod nextpc;
 mod opcodes;
 pub mod registers;
+pub mod get_at;
+pub mod reader;
 
 pub use crate::bus::RegisterBus;
 pub use crate::cpu::Cpu;
 pub use crate::interface::{NewRegisters, Registers};
 pub(crate) use flags::Flags;
+pub(crate) use crate::reader::Reader;
